+++ crates/core/src/drive/sync.rs
@@ -642,7 +642,6 @@ mod tests {
                 total_pages: Some(200),
                 finished_at: None,
                 last_read_at: "2026-08-23 09:00:00".into(),
-                scroll_position: 0.0,
             },
         )
         .unwrap();
@@ -726,7 +725,6 @@ mod tests {
                 total_pages: Some(200),
                 finished_at: None,
                 last_read_at: "2026-08-23 09:00:00".into(),
-                scroll_position: 0.0,
             },
         )
         .unwrap();
@@ -778,7 +776,6 @@ mod tests {
                 total_pages: Some(200),
                 finished_at: None,
                 last_read_at: "2026-08-23 09:00:00".into(),
-                scroll_position: 0.0,
             },
         )
         .unwrap();
@@ -828,7 +825,6 @@ mod tests {
                 total_pages: Some(200),
                 finished_at: None,
                 last_read_at: "2026-08-23 10:00:00".into(),
-                scroll_position: 0.0,
             },
         )
         .unwrap();
@@ -887,7 +883,6 @@ mod tests {
                 total_pages: Some(200),
                 finished_at: None,
                 last_read_at: "2026-08-23 09:00:00".into(),
-                scroll_position: 0.0,
             },
         )
         .unwrap();